Introducing Slack transport
