--- tests/resources/test-config-scan-log-console.json
@@ -1,6 +1,0 @@
-{
-  "target": "192.168.0.2",
-  "logger": {
-    "transport": "console"
-  }
-}
