+++ tests/resources/test-config-scan.json
@@ -1,3 +1,6 @@
 {
-    "target": "192.168.0.2"
+  "target": "192.168.0.2",
+  "logger": {
+    "transport": "console"
+  }
 }
